+++ src/qler/__init__.py
@@ -24,7 +24,7 @@ from qler.queue import Queue
 from qler.task import task
 from qler.worker import Worker
 
-__version__ = "0.5.0"
+__version__ = "0.5.1"
 
 __all__ = [
     "__version__",